Add check if types already contains path
This commit updates `World::declare` to check if the types hashmap
already contains an entry for the passed-in `path` argument.

The motivation for adding this check is that without it, it is possible
for the `type_slots` vector to contain entries that are "orphaned",
meaning that they are not pointed to by the `types` hashmap (the value
of this hashmap is an index into the type_slots vector.
For more details about this issue please see the referenced writeup
below.

Refs: https://github.com/danbev/learning-crypto/blob/main/notes/seedwing/playground-evaulate-issue.md
Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>