Accept YAML input and introduce 'format' query param
Fixes #64 which is arguably not broken ;)

We've moved back to a single `evaluate` handler that will accept
either JSON or YAML input -- if it fails one, it tries the other --
and could be easily extended to support TOML and other formats.

By default, the response will in the format indicated by the
Content-Type header of the request, i.e. if you pass in JSON, you get
JSON back.

If you pass in an unrecognized Content-Type,
e.g. application/x-www-form-urlencoded, you'll get HTML back.

You can override this default behavior to control the format of the
response by including the `format` query param in the request, valid
values being `html`, `json`, or `yaml`.

Signed-off-by: Jim Crossley <jim@crossleys.org>