Clone data_sources in World::clone
This commit add cloning of the data_sources in World::clone.

The motivation for this is what without cloning the data_sources
data::from usages will not work in the playground.

For example, currently the example in the documentation which uses
data::from does not work:
```
pattern people = lang::or<*data::from<"people.json">>
```

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>