Add error logging when target is not found
This commit adds an error log when the target cannot be found.

The motivation for this is that currently it is easy to make a typo in
the file name of a `data::from`, in which case the rule will fail but
there is no information indicating that the file name could be
incorrect.  Having this logging helps but long term we should probably
Err on this situation.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>