fix: rework frontend build to mitigate extraneous trunk builds
These were mostly caused by a missing yarn.lock file, which could
occur if `npm i` is used, as is the case for the CI workflow.

Moving on from yarn in favor of `npm ci` which guarantees repeatable
builds by reading the lock file to recreate node_modules/

We first attempted to remove filename hash suffixes from the trunk
output. This worked but it could cause caching errors in the browser.

We then reverted that and tried building the assets beneath
OUT_DIR. This seemed to work!

We also now use the npm lock file instead of package.json to determine
when a build should re-run. The package.json isn't enough as it won't
necessarily refer to exact versions like the -lock file.

Signed-off-by: Jim Crossley <jim@crossleys.org>