refactor(server): change the URL structure a bit
This removes the mandatory addition of the trailing slash. Seemed good
in the beginning, then cause some issues.

This also mounts the OpenAPI spec under /api, so that it is different
from the '/openapi' prefix with the embedded UI, which caused some
issues with trunk.

Finally, is proxies the swaggerui application with trunk.