refactor(engine): pull out the mir->lir lowering stuff
This moves the logic lowering mir->lir into the a dedicated struct.
Making it easier to work with the data in a more complex setup, when
adding packages.

This also gives the runtime::World a "ready-to-run" state, and makes it
immutable, as further changes on the world should go through the chain
or hir->mir->lir.

With that change, it is possible to generate a type hierarchy, and
serve requests from the existing state, rather than building the
package hierarchy with every request.