feat: add test_patterns function and test utils
This commit adds a test_patterns function which can be used when there
are multiple patterns to be tested.

This commit also contains the adding the directory `engine/test-data`
as a data source directory to the builder to enable functions like
`data::from` to be used in tests. For example:
```
  pattern blob = *data::from<"intoto/binary-linux-amd64">
```

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>