fix: prevent the UI from throwing errors for simple schemas
When using the Schema "bool" (which just means, match anything, or
nothing) the Swagger UI will throw an exception, as it is not supported.

We did capture this on the outer layer, but not nested. This change
switches to use the Rust type for that complex schema (not bool),
ensuring that we don't run into this issue, converting where necessary.

Also, this change groups by the first element of the path.