test: refactor test to be offline only
The proxy test requires access to an external service. And it's data
might change over time.

So the proxy test is "ignored", we can still run it manually, but the
actual test uses the serialized response data, as the part we wanted to
test was the collapsing of the response.