refactor: extract collect from collapse
This change extracts the core part of the 'collapse' method,
gathering all failed leaves into a dedicated 'collect' method which
doesn't return the root node either.

The collapse method still works the same way, but is now built on top
of the collect method.

The collapse method had the disadvantage that it always returned the
root node, as it requires the 'rationale' field to return a list.