feat: return json body for NotAcceptable response
This commit contains a suggestion to make evaluate return a json body
instead of a string which is currently the case.

The motivation for this is that there may be situations where the
pattern does not compile and this feedback might be useful to the writer
of the pattern. One example of this is where a pattern contains a
duplicate field which would be reported back in the playground with this
the changes in this commit.

With that said, I'm sure there are cleaner ways to add this to the code
base but I wanted to get some feedback on this before moving forward.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>