Set release to true to Trunk.toml
This commit suggests setting the trunk configuration option `release`
to true (default is false) in Trunk.toml.

The motivation for this is that a release build will invoke [wasm-opt]
which performs a number of [optimizations] to make wasm smaller and
faster.

This could also be specified as a environment variable if setting this
in the Trunk.toml configuration file is the best option:
```
TRUNK_BUILD_RELEASE=true
```

[wasm-opt]: https://crates.io/crates/wasm-opt
[optimizations]: https://github.com/WebAssembly/binaryen#binaryen-optimizations

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>