feat: add initial static component executor
This commit adds a composed webassembly component model module that uses
the policy engine module.

The idea here is that we can take a policy, perhaps created in the
playground, and include it in the static component. This component
will import the policy engine and call the eval function in the policy
engine component module. This way a user can have a self contained
.wasm which can be executed by and wasm runtime that supports the
webassembly component model.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>